add stacking visualisation
