+++ client/components/pages/order/dispatchUser/StackOptions.tsx
@@ -29,60 +29,12 @@ import { Switch, Box, ButtonGroup, Button } from "@mui/material";
 
 import * as L from "leaflet";
 
-import SVGhomeFlag from "../../../../static/icons/house-flag-solid-green.svg";
-import SVGhomeFlagBlue from "../../../../static/icons/house-flag-solid-blue.svg";
-import SVGUserSolid from "../../../../static/icons/user-solid.svg";
+import {outletIcon,endLocationIcon,currentLocationIcon} from "./Iconz"
 
-interface BoundSetterProps {
-  currentLocation: Coordinate;
-  stackEndLocation: Coordinate;
-  outletLocation: Coordinate;
-}
-const BoundSetter: React.FC<BoundSetterProps> = ({
-  outletLocation,
-  currentLocation,
-  stackEndLocation,
-}) => {
-  const map = useMap();
 
-  React.useEffect(() => {
-    console.log(`[Boundsetter useeffect`);
-    const coords = [currentLocation, stackEndLocation, outletLocation]
-      .filter((c) => !!c)
-      .map((c) => L.marker(c));
+import BoundSetter from "./BoundSetter"
 
-    if (coords.length > 1) {
-      map.fitBounds(L.featureGroup(coords).getBounds());
-      console.log(`zoom level ${map.getZoom()}`);
-    } else {
-      map.setView(coords[0].getLatLng(), 15);
-    }
-  }, [
-    currentLocation[0],
-    currentLocation[1],
-    stackEndLocation[0],
-    stackEndLocation[1],
-    outletLocation[0],
-    outletLocation[1],
-  ]);
 
-  return <></>;
-};
-
-const outletIcon = L.icon({
-  iconUrl: SVGhomeFlag,
-  iconSize: [10, 10],
-});
-
-const endLocationIcon = L.icon({
-  iconUrl: SVGhomeFlagBlue,
-  iconSize: [10, 10],
-});
-
-const currentLocationIcon = L.icon({
-  iconUrl: SVGUserSolid,
-  iconSize: [10, 10],
-});
 
 interface StackOptionsProps {
   stackWindow: number;
